server/requirements.txt
grpcio==1.60.0
pyvisa==1.13.0